Web components (initial implementation)
Shifts the interface from a react component / hook to a web component.
The web component only has a dependency on d3, and removing React makes
the code much simpler as we can write the code in a simple imperative
style.

There's a number of aspects not implemented here, including:

* an attribute-based interface to configure the component
* bundling / distribution of the component
* distributing CSS via the component itself
* options UI, styling, panel layout etc
* the drag-and-drop page surfaced via GitHub pages
