+++ index.html
@@ -1,0 +1,30 @@
+<!doctype html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>evofr | Web Components</title>
+  <link rel="icon" type="image/x-icon" href="favicon.ico">
+  <link rel="stylesheet" href="styles.css">
+
+  <!-- Following styles should be part of the component itself -->
+  <link rel="stylesheet" href="src/lib/styles/styles.css">
+  <link rel="stylesheet" href="src/lib/styles/rc-switch.css">
+</head>
+
+<body>
+  <main>
+    <h1>evofr | Web Components</h1>
+    <br/>
+    
+    <evofr-viz>
+      <evofr-controls logit="1"/>
+    </evofr-viz>
+
+
+  </main>
+  <script type="module" src="lib/index.ts"></script>
+</body>
+
+</html>
