+++ lib/cssSafeName.js
