+++ lib/d3Graph.js
@@ -1,12 +1,13 @@
 import * as d3 from "d3";
-import { logitScale } from "./logitScale";
-import { Tooltip } from "./tooltip";
-import { cssSafeName } from "./cssSafeName";
+import { logitScale } from "./logitScale.js";
+import { Tooltip } from "./tooltip.js";
+import { cssSafeName } from "./cssSafeName.js";
 
 const TRANSITION_DURATION = 700;
 
 export function D3Graph(d3Container, sizes, modelData, params, options) {
-  const dom = d3.select(d3Container.current);
+  // const dom = d3.select(d3Container.current);
+  const dom = d3.select(d3Container);
   this.svg = svgSetup(dom, sizes);
   this.tooltip = new Tooltip(dom);
   this.modelData = modelData;
