+++ lib/parseModelData.js
