+++ lib/tooltip.js
@@ -123,4 +123,3 @@ export class Tooltip {
     }
   }
 }
-
