+++ lib/tooltipDisplay.js
