+++ styles.css
@@ -1,0 +1,86 @@
+:root {
+  font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
+  line-height: 1.5;
+  font-weight: 400;
+
+  color: #213547;
+  background-color: #ffffff;
+
+  font-synthesis: none;
+  text-rendering: optimizeLegibility;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+
+  --text-color: #535359;
+}
+
+* {
+  box-sizing: border-box;
+}
+
+html,
+body,
+main {
+  min-height: 100vh;
+  min-height: 100svh;
+}
+
+body {
+  margin: 0;
+  padding: 0;
+  font-size: 1.1rem;
+  background: #fafafa;
+  color: #333;
+
+  @media (min-width: 600px) {
+    padding: 0 2rem;
+  }
+}
+
+main {
+  max-width: 1000px;
+  margin: 0 auto;
+  background: white;
+  padding: 2rem;
+  padding-bottom: 4rem;
+
+  @media (min-width: 600px) {
+    padding: 4rem;
+    padding-bottom: 6rem;
+    border-left: 1px solid #f4f4f6;
+    border-right: 1px solid #f4f4f6;
+  }
+}
+
+h1,
+h2 {
+  margin: 0;
+  line-height: 1.2;
+}
+
+h2 {
+  margin: 3rem 0 1rem;
+}
+
+p {
+  color: var(--text-color);
+  max-width: 25em;
+}
+
+a {
+  color: #0070f3;
+  text-decoration: none;
+  transition: color 0.2s ease;
+  padding: 0;
+  margin: 0;
+  font: inherit;
+  background: none;
+  border: none;
+  cursor: pointer;
+  text-decoration: underline;
+  text-decoration-color: #ccc;
+
+  &:hover {
+    text-decoration-color: currentColor;
+  }
+}
